Refactor top-level code and Game class for test.
The test of create_players was proving tricky, perhaps because as
soon as the Game object was created, it was calling a chain of
methods including create_players. I wanted to 'allow' the game
object to receive ask_first_name and ask_second_name to supply
'Peter' and 'Chris' but the object didn't exist yet.
So my best idea was to put play_game in the top-level code in
play_connect_four so that the Game object can 'sit there' with
nothing happening before the play_game method is called.
This allows the test of create_players to work.